+++ example/src/components/bind-passkey.tsx
@@ -2,15 +2,15 @@ import { useState } from "react";
 import "bootstrap/dist/css/bootstrap.css";
 import Highlight from "react-highlight";
 
-const BindCredential = () => {
-  const [bindCredentialUsername, setBindCredentialUsername] = useState(String);
-  const [bindCredentialResult, setBindCredentialResult] = useState({});
+const BindPasskey = () => {
+  const [bindPasskeyUsername, setBindPasskeyUsername] = useState(String);
+  const [bindPasskeyResult, setBindPasskeyResult] = useState({});
 
-  async function handleBindCredentialClick(e: React.MouseEvent<HTMLButtonElement>) {
+  async function handleBindPasskeyClick(e: React.MouseEvent<HTMLButtonElement>) {
     e.preventDefault();
     const BeyondIdentityEmbeddedSdk = await import("../utils/BeyondIdentityEmbeddedSdk");
     let embedded = new BeyondIdentityEmbeddedSdk.default();
-    let username = bindCredentialUsername;
+    let username = bindPasskeyUsername;
     let response = await fetch('https://acme-cloud.byndid.com/credential-binding-link', {
       method: 'POST',
       headers: {
@@ -24,16 +24,16 @@ const BindCredential = () => {
     });
     let jsonResponse = await response.json();
     if (response.status !== 200 || jsonResponse === null) {
-      setBindCredentialResult(jsonResponse);
+      setBindPasskeyResult(jsonResponse);
       return;
     }
     let credentialBindingLink = jsonResponse.credential_binding_link;
-    if (await embedded.isBindCredentialUrl(credentialBindingLink)) {
-      let result = await embedded.bindCredential(credentialBindingLink);
-      setBindCredentialResult(result);
-      window.postMessage("update-credentials", "*");
+    if (await embedded.isBindPasskeyUrl(credentialBindingLink)) {
+      let result = await embedded.bindPasskey(credentialBindingLink);
+      setBindPasskeyResult(result);
+      window.postMessage("update-passkeys", "*");
     } else {
-      setBindCredentialResult(jsonResponse);
+      setBindPasskeyResult(jsonResponse);
     }
   }
 
@@ -55,13 +55,13 @@ const BindCredential = () => {
                     id="bindCredInput"
                     type="text"
                     className="form-control rounded-4 mb-3"
-                    onChange={event => setBindCredentialUsername(event.target.value)}
+                    onChange={event => setBindPasskeyUsername(event.target.value)}
                   />
-                  <label htmlFor="bindCredentialUsername">Username</label>
+                  <label htmlFor="bindPasskeyUsername">Username</label>
                   <button
                     id="bindCredButton"
                     type="button"
-                    onClick={handleBindCredentialClick}
+                    onClick={handleBindPasskeyClick}
                     className="btn btn-primary btn-lg px-4"
                   >
                     Create
@@ -70,11 +70,11 @@ const BindCredential = () => {
               </div>
             </div>
 
-            {Object.keys(bindCredentialResult).length > 0 &&
+            {Object.keys(bindPasskeyResult).length > 0 &&
               <div className="row row-cols-1 row-cols-md-1 mt-3">
                 <div id="bindCredResultDiv" className="col">
                   <Highlight className='json'>
-                    {JSON.stringify(bindCredentialResult, null, 2)}
+                    {JSON.stringify(bindPasskeyResult, null, 2)}
                   </Highlight>
                 </div>
               </div>
@@ -86,4 +86,4 @@ const BindCredential = () => {
   );
 };
 
-export default BindCredential;
+export default BindPasskey;