+++ example/src/components/header.tsx
@@ -1,6 +1,5 @@
 import styles from "./header.module.css"
-import Link from "next/link"
-import { signIn, signOut, useSession } from "next-auth/react"
+import { signOut, useSession } from "next-auth/react"
 
 export default function Header() {