+++ example/src/pages/api/auth/[...nextauth].ts
@@ -1,4 +1,6 @@
 import NextAuth, { NextAuthOptions } from "next-auth"
+import Auth0Provider from "next-auth/providers/auth0"
+import OktaProvider from "next-auth/providers/okta"
 
 // For more information on each option (and a full list of options) go to
 // https://next-auth.js.org/configuration/options
@@ -23,6 +25,27 @@ export const authOptions: NextAuthOptions = {
         }
       },
     },
+    Auth0Provider({
+      clientId: "q1cubQfeZWnajq5YkeZVD3NauRqU4vNs",
+      clientSecret: "2NXvQo_WYo7PeLoutxGyhNryhBWK807wMuibBfxZvHH6e5ZGK7k0glQUZUBLdufU",
+      authorization: "https://dev-pt10fbkg.us.auth0.com/authorize?connection=Example-App-Web",
+      issuer: "https://dev-pt10fbkg.us.auth0.com",
+      profile(profile) {
+        let split = profile.sub.split("|");
+        let email = split[split.length - 1];
+        return {
+          id: email,
+          name: email,
+          email: email,
+        }
+      },
+    }),
+    OktaProvider({
+      clientId: "0oa6xervrvbPQ1F4H5d7",
+      clientSecret: "xPUMnQa3Vt0oDADtIV6YmJvYe3WwuoapxU5jtYFd",
+      authorization: "https://dev-43409302.okta.com/oauth2/v1/authorize?idp=0oa5rszht04BEucVQ5d7",
+      issuer: "https://dev-43409302.okta.com",
+    }),
   ],
   theme: {
     colorScheme: "light",