+++ src/index.ts
@@ -1,10 +1,8 @@
 export {
-  CredentialV1 as Credential,
   TenantId,
   RealmId,
   IdentityId,
-  CredentialId,
   Log,
   ConsoleLog,
 } from "coresdk";
-export { Embedded } from "./embedded";
+export { Embedded, Passkey } from "./embedded";