+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-01-09
+### Added
+- Support for authenticating with Auth0 and Okta in the example app
+
+### Changed
+- Rename instances of `Credential` to `Passkey`
+- Nest tenantId, realmId, and identityId under appropriate objects in the `Credential`
+- Update support links in the example app
+
+### Fixed
+- Scheme without a path is now recognized as a valid URL when binding a credential
+- Issues where SWC wasn't able to compile the JS SDK
+
 ## [1.0.10] - 2022-10-17
 
 ### Changed